+++ src/server/Model/Search/IDatasetCompilerContextProvider.cs
@@ -4,11 +4,15 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 using System;
+using System.Threading;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Model.Compiler;
 
-namespace Model.Compiler
+namespace Model.Search
 {
-    public interface IPanelValidator
+    public interface IDatasetCompilerContextProvider
     {
-        PatientCountQuery Validate(PanelValidationContext ctx);
+        Task<DatasetCompilerContext> GetCompilerContextAsync(DatasetExecutionRequest request);
     }
 }
